Make camera mute work with proprietary libcameraservice
This makes the persist.sys.camera-mute and ro.camera.sound.disabled
properties work with the proprietary libcameraservice.so.

This is done by first patching the libcameraservice.so during build
to not use the .ogg files in /system/media/audio/ui but rather in
/data/system/camera. /data/system/camera contains symlinks pointing to
either /dev/null if the shutter sound is to be muted or to the actual
ogg files.

The symlinks in /data/system/camera are maintained by the
"toggleshutter" program. It is started by init via the "on property"
trigger mechanism and checks whether the symlinks reflect the
properties' state and, if not, updates them accordingly.

To use this, use libcameraservice.patched.so rather than
libcameraservice.so in your PRODUCT_COPY_FILES definition and add
toggleshutter to PRODUCT_PACKAGES.

The libcameraservice.patched.so is currently created in the same
directory as the libcameraservice.so
(vendor/samsung/aries-common/proprietary), this should be fixed. It's
just that I don't know how to make this better, even after struggling
with the build system for hours.

Change-Id: I031461c9366f276a84c7eb72b629ceb45e8c1e92